9540: add check to existence of judge on trial session, update and split tests into two files, remove old comments
