+++ shared/src/business/useCases/trialSessions/updateTrialSessionInteractor.js
@@ -73,10 +73,13 @@ const updateAssociatedCaseAndSetNoticeOfChange = async ({
         });
     }
 
+    // should issue notice of change of trial judge only when changing from one judge to another
     const shouldIssueNoticeOfChangeOfTrialJudge =
       currentTrialSession.isCalendared &&
-      currentTrialSession.judge?.userId !==
-        updatedTrialSessionEntity.judge?.userId &&
+      currentTrialSession.judge &&
+      updatedTrialSessionEntity.judge &&
+      currentTrialSession.judge.userId !==
+        updatedTrialSessionEntity.judge.userId &&
       caseEntity.status !== CASE_STATUS_TYPES.closed;
 
     if (shouldIssueNoticeOfChangeOfTrialJudge) {
